feat: 테마 삭제 시 예약 대기 존재 검증 기능 추가

+++ src/main/java/roomescape/service/ThemeService.java
@@ -12,6 +12,7 @@ import roomescape.common.exception.ConflictException;
 import roomescape.common.exception.NotFoundException;
 import roomescape.dao.ReservationDao;
 import roomescape.dao.ThemeDao;
+import roomescape.dao.WaitingDao;
 import roomescape.dao.dto.TimeQueryResult;
 import roomescape.domain.reservation.theme.Description;
 import roomescape.domain.reservation.theme.Theme;
@@ -26,11 +27,13 @@ public class ThemeService {
     private static final String uploadDir = System.getProperty("user.dir") + "/src/main/resources/static/images/";
 
     private final ThemeDao themeDao;
+    private final WaitingDao waitingDao;
     private final ReservationDao reservationDao;
     private final Clock clock;
 
-    public ThemeService(ThemeDao themeDao, ReservationDao reservationDao, Clock clock) {
+    public ThemeService(ThemeDao themeDao, WaitingDao waitingDao, ReservationDao reservationDao, Clock clock) {
         this.themeDao = themeDao;
+        this.waitingDao = waitingDao;
         this.reservationDao = reservationDao;
         this.clock = clock;
     }
@@ -90,6 +93,10 @@ public class ThemeService {
 
         if (reservationDao.existsByThemeId(id)) {
             throw new ConflictException("예약이 존재하는 테마는 삭제할 수 없습니다.");
+        }
+
+        if (waitingDao.existsByThemeId(id)) {
+            throw new ConflictException("예약 대기가 존재하는 테마는 삭제할 수 없습니다.");
         }
 
         themeDao.delete(id);
